pyproject.toml
[project]
name = "ai-discord-chatbot-tutorial"
version = "0.1.0"
description = "Serverless AI Discord chatbot tutorial using Amazon Bedrock"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.36.16",
    "pynacl>=1.5.0",
    "requests>=2.32.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.taskipy.tasks]
export = "uv pip compile pyproject.toml > requirements.txt"
